requirements.txt
# Core dependencies
streamlit==1.32.0
plotly==5.18.0
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
alembic==1.13.1
psycopg2-binary==2.9.9
chromadb==0.3.29
langchain==0.1.9
openai==1.12.0
numpy==1.24.3
python-dotenv==1.0.1
tenacity==8.2.3
tqdm==4.66.2
requests==2.31.0
reportlab==4.0.4
kaleido==0.2.1
bcrypt==4.1.2

# Optional dependencies for production
streamlit-webrtc==0.47.1
python-docx==1.1.0
pdf2image==1.17.0
pytesseract==0.3.10
PyMuPDF==1.23.8
docx2txt==0.8
SpeechRecognition==3.10.0
gTTS==2.3.2
openai-whisper==20231117
soundfile==1.12.1
spacy==3.7.4 